backend/requirements.txt
fastapi
uvicorn
python-jose
httpx                     
redis                     
python-dotenv              
